# chunks: decode varints directly from byte buffer; stop panicing on some corrupt inputs