--- a/pkg/chunkenc/memchunk.go
+++ b/pkg/chunkenc/memchunk.go
@@ -1146,7 +1146,12 @@ func (si *bufferedIterator) Next() bool {
 
 	if !si.closed && si.reader == nil {
 		// initialize reader now, hopefully reusing one of the previous readers
-		si.reader = si.pool.GetReader(bytes.NewBuffer(si.origBytes))
+		var err error
+		si.reader, err = si.pool.GetReader(bytes.NewBuffer(si.origBytes))
+		if err != nil {
+			si.err = err
+			return false
+		}
 	}
 
 	ts, line, ok := si.moveNext()