@bboreham 2022-09-29
I added a test for the infinite loop that I foresaw, and that made the gzip decoder panic, so I fixed both issues.